+++ feature/login/src/main/java/com/example/login/auth/EmailAuthViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.linku_android.auth
+package com.example.login.auth
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
