+++ feature/login/src/main/java/com/example/login/auth/SignUpPasswordScreen.kt
@@ -1,4 +1,4 @@
-package com.example.linku_android.auth
+package com.example.login.auth
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
